Replace wxMin / wxMax / wxClip with std::min, std::max, and std::clamp. Avoid implicit conversions.
Long is really becoming annoying, and should probably be removed in favor of exact sign / size integers (i.e. int64_t, int32_t etc.)

+++ include/wx/generic/grid.h
@@ -1636,7 +1636,7 @@ public:
 
     void GetTextBoxSize( const wxDC& dc,
                          const std::vector<std::string>& lines,
-                         long *width, long *height ) const;
+                         int* width, int* height ) const;
 
     // If bottomRight is invalid, i.e. == wxGridNoCellCoords, it defaults to
     // topLeft. If topLeft itself is invalid, the function simply returns.
